+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/SDK.java
@@ -379,6 +379,9 @@ public class SDK {
         req.setMethod("PUT");
         req.setURL(url);
         SerializedBody serializedRequestBody = org.openapis.openapi.utils.Utils.serializeRequestBody(request, "request", "string");
+        if (serializedRequestBody == null) {
+            throw new Exception("Request body is required");
+        }
         req.setBody(serializedRequestBody);
 
         req.addHeader("Accept", "application/json");