+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/SDKConfiguration.java
@@ -35,7 +35,7 @@ class SDKConfiguration {
 	}};
 	public String language = "java";
 	public String openapiDocVersion = "0.1.0";
-	public String sdkVersion = "1.31.8";
+	public String sdkVersion = "1.31.9";
 	public String genVersion = "2.115.2";
 	public Map<String, Map<String, Map<String, Object>>> globals = new HashMap<String, Map<String, Map<String, Object>>>(){{
 		put("parameters", new HashMap<String, Map<String, Object>>());