+++ php-client-sdk/src/Resource.php
@@ -69,7 +69,7 @@ class Resource
      * @return \OpenAPI\OpenAPI\Models\Operations\CreateResourceResponse
      */
 	public function createResource(
-        ?\OpenAPI\OpenAPI\Models\Shared\ExampleResource $request,
+        \OpenAPI\OpenAPI\Models\Shared\ExampleResource $request,
     ): \OpenAPI\OpenAPI\Models\Operations\CreateResourceResponse
     {
         $baseUrl = Utils\Utils::templateUrl($this->sdkConfiguration->getServerUrl(), $this->sdkConfiguration->getServerDefaults());
@@ -77,9 +77,10 @@ class Resource
         
         $options = ['http_errors' => false];
         $body = Utils\Utils::serializeRequestBody($request, "request", "json");
-        if ($body !== null) {
-            $options = array_merge_recursive($options, $body);
+        if ($body === null) {
+            throw new \Exception('Request body is required');
         }
+        $options = array_merge_recursive($options, $body);
         $options['headers']['Accept'] = 'application/json';
         $options['headers']['x-speakeasy-user-agent'] = sprintf('speakeasy-sdk/%s %s %s %s', $this->sdkConfiguration->language, $this->sdkConfiguration->sdkVersion, $this->sdkConfiguration->genVersion, $this->sdkConfiguration->openapiDocVersion);