+++ typescript-client-sdk/src/sdk/resource.ts
@@ -141,6 +141,8 @@ export class Resource {
         }
         const properties = utils.parseSecurityProperties(globalSecurity);
         const headers = { ...reqBodyHeaders, ...config?.headers, ...properties.headers };
+        if (reqBody == null || Object.keys(reqBody).length === 0)
+            throw new Error("request body is required");
         headers["Accept"] = "application/json";
 
         headers[